refactor(renderer): move the lyric dom builder into the module
