+++ src/renderer/seek.ts
